back_end/requirements.txt
djangorestframework
markdown
django-filter
django-cors-headers
django
drf-yasg
drf_excel
django-storages
django-environ
google-cloud-storage
psycopg2
drf-nested-routers
sentry-sdk[django]
shortuuid
channels
openai
redis
Celery
celery[redis]
celery[django]
django-celery-results
django_redis
django-firebase-auth>=1.2.1
pillow